Improve the grpc command to only create essentials client and service modes
