+++ stdlib/grpc/src/main/resources/templates/skeleton/service_sample.mustache
@@ -38,3 +38,12 @@ service {{serviceName}} on ep {
         // You should return a {{outputType}}
     }{{/unaryFunctions}}
 }
+
+{{#each messageList}}{{> message}}{{/each}}
+{{#enumList}}{{> enum}}{{/enumList}}{{#if rootDescriptor}}const string ROOT_DESCRIPTOR = "{{rootDescriptor}}";
+function getDescriptorMap() returns map<string> {
+    return {
+        {{#each descriptors}}"{{descriptorKey}}":"{{descriptorData}}"{{#unless @last}},{{/unless}}
+        {{/each}}
+    };
+}{{/if}}
